+++ tests/Protocol/MessageTest.php
@@ -1,5 +1,5 @@
 <?php
-namespace ParagonIE\Gossamer\Tests;
+namespace ParagonIE\Gossamer\Tests\Protocol;
 
 use PHPUnit\Framework\TestCase;
 use ParagonIE\Gossamer\Protocol\Message;
@@ -8,6 +8,7 @@ use SodiumException;
 
 /**
  * Class MessageTest
+ * @package ParagonIE\Gossamer\Tests\Protocol
  */
 class MessageTest extends TestCase
 {
