+++ tests/Protocol/PacketTest.php
@@ -1,5 +1,5 @@
 <?php
-namespace ParagonIE\Gossamer\Tests;
+namespace ParagonIE\Gossamer\Tests\Protocol;
 
 use ParagonIE\Gossamer\Protocol\Packet;
 use PHPUnit\Framework\TestCase;
@@ -8,6 +8,7 @@ use SodiumException;
 
 /**
  * Class PacketTest
+ * @package ParagonIE\Gossamer\Tests\Protocol
  */
 class PacketTest extends TestCase
 {
